+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.1.0 (in progress)
+
++ Add version constraint compare function.
++ Add generic iterator class.
++ Added require-dev section reader
+
 ## 1.0.1 (24. September 2018)
 
 + Add continuous testing integration.